separate plugins.rs into multiple files + rename things
