+++ fud2/fud-core/src/lib.rs
@@ -1,9 +1,9 @@
 pub mod cli;
 pub mod config;
 pub mod exec;
-pub mod plugins;
 pub mod run;
+pub mod script;
 pub mod utils;
 
 pub use exec::{Driver, DriverBuilder};
-pub use plugins::LoadPlugins;
+pub use script::LoadPlugins;
