+++ fud2/fud-core/src/script/mod.rs
@@ -1,0 +1,5 @@
+mod exec_scripts;
+mod plugin;
+mod report;
+
+pub use plugin::LoadPlugins;
